Per extruder config, retract jerk, extruder offset - per extruder config   Major parameters like steps/mm, acceleration, etc. are now configurable   for each extruder. - retract jerk   Added DEFAULT_E_JERK parameter for extruder retract moves to allow fine   tuning (without it the intial speed varies and retract is not efficient   in eliminationg blobs) - extruder offset   Added parameter for extruder offset that allows configuring how the   priniting head should be repositioned when switching extruders.   The T* command can now also accept the feedrate at which to reposition   the head when switching extruders. Notes:   There were changes in LCD management code I did not test.   In general, the commands and settings for extruder are applied only to the   active extruder. For example, to change max feedrate for extruders 0 and   1, one has to switch to the first extruder, set its rate, then switch to   the second and set its rate. M503 shows multiple 'E' values starting   with extruder 0 and so on.   As of this commit the firmware can be used to print suport in different   material by simply switching extruder in skeinforge support start/end   script and adjusting the steps/mm, etc. to make the same gcode work for both   main and support material extruders.

+++ Marlin/Marlin.h
@@ -58,7 +58,7 @@ const prog_char echomagic[] PROGMEM ="echo:";
 #define SERIAL_ECHOPAIR(name,value) {SERIAL_ECHOPGM(name);SERIAL_ECHO(value);}
 
 // Macro for getting current active extruder
-#define ACTIVE_EXTRUDER (active_extruder)
+#define ACTIVE_EXTRUDER ((int)active_extruder)
 
 //things to write to serial from Programmemory. saves 400 to 2k of RAM.
 #define SerialprintPGM(x) serialprintPGM(MYPGM(x))
@@ -147,7 +147,7 @@ void prepare_arc_move(char isclockwise);
 
 extern float homing_feedrate[];
 extern bool axis_relative_modes[];
-extern float current_position[NUM_AXIS] ;
+extern float current_position[NUM_AXIS];
 extern float add_homeing[3];
 
 // Handling multiple extruders pins
